+++ modules/main-chats/init.js
