+++ modules/tasks/helpers/task_class_Mongo.js
@@ -14,7 +14,7 @@ const { handleTaskAttachmentsDuplicateFunctionality } = require(`../../../common
 const { buildQueryObject, buildHistoryObject, convertToDisplayFormat } = require("./helper");
 const socketEmitter = require('../../../event/socketEventEmitter');
 const { addCommentCollection, updateCommentCollection } = require('../../comments/controller')
-const { updateMainChat } = require('../../MainChats/controller');
+const { updateMainChat } = require('../../main-chats/controller');
 const { replaceObjectKey } = require("../../auth/helper");
 const { emitListener } = require("../../company/eventController.js");
 const { createCustomFields } = require("./helper.js");
